launch: Change launch/linux/sev module to launch/linux module
